+++ pkg/providers/gtm/testdata/TestDataGtmCidrmap/missing_map_name.tf
@@ -1,0 +1,7 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_gtm_cidrmap" "gtm_cidrmap" {
+  domain = "test.cidrmap.domain.net"
+}