+++ CHANGELOG.md
@@ -77,10 +77,13 @@
 * GTM
     * Added data sources:
         * `akamai_gtm_domains` - reads list of GTM domains under a given contract.
+
+* GTM
+    * Added data sources:
+        * `akamai_gtm_cidrmap` - reads information for a specific GTM cidrmap    
     
 * PAPI
   * Improve errors for `akamai_contract` and `akamai_group` datasources when there are multiple groups or contracts.
-